requirements.txt
Flask
requests
psycopg2-binary
flask-login
werkzeug
python-dotenv
pytz

